# Added support for bulk insert of Notes: 2000% faster