@marcardar 2016-03-02
Thanks very much. I've just made some significant changes to the PR. After thinking about this initial approach, I decided using the deck id in the content values for notes was a bit of a hack (not very consistent with the way content values are currently used in the provider). Instead, the content provider now supports a "deckId" query parameter.

When inserting multiple notes (bulkInsert()) this parameter directs the provider to insert all cards into that deck. If the parameter is not included, then we get the old behaviour.

Similarly, when inserting a single note, the deckId parameter instructs the provider to insert cards into the deck. If the parameter is not included, then we get the old behaviour (which is to insert the note and return the uri without adding cards to a deck).