@timrae 2016-03-07
> Added API support for counting number of notes (with a certain model id) in a deck: countNotes()

What are you using this for?