@timrae 2016-03-07
Would the deck ID query parameter just be ignored by version 1 of the provider? If so you could heavily simplify the code to this:

``` java
public Uri addNewNote(long mid, long did, String[] fields, String tags) {
    ContentValues values = new ContentValues();
    values.put(FlashCardsContract.Note.MID, mid);
    values.put(FlashCardsContract.Note.FLDS, joinFields(fields));
    values.put(FlashCardsContract.Note.TAGS, tags);
    Uri.Builder builder = FlashCardsContract.Note.CONTENT_URI.buildUpon();
    Uri uri = builder.appendQueryParameter(FlashCardsContract.Note.DECK_ID_QUERY_PARAM, String.valueOf(did)).build();
    Uri newNoteUri = mResolver.insert(uri, values);
    if (newNoteUri != null && getProviderSpecVersionCode() < DECK_ID_QUERY_PARAM_MIN_SPEC_VERSION) {
        // Version 2+ of the provider sets the deck automatically so this is only necessary for spec version 1
        moveNewCardsToDeck(newNoteUri, did);
    }
    return newNoteUri;
}

private void moveNewCardsToDeck(Uri newNoteUri, long did) {
    // Move cards to specified deck (not necessary from spec version 2 of the content provider)
    Uri cardsUri = Uri.withAppendedPath(newNoteUri, "cards");
    final Cursor cardsCursor = mResolver.query(cardsUri, null, null, null, null);
    try {
        ContentValues cardValues = new ContentValues();
        cardValues.put(FlashCardsContract.Card.DECK_ID, did);
        while (cardsCursor.moveToNext()) {
            String ord = cardsCursor.getString(cardsCursor.getColumnIndex(FlashCardsContract.Card.CARD_ORD));
            Uri cardUri = Uri.withAppendedPath(cardsUri, ord);
            mResolver.update(cardUri, cardValues, null, null);
        }
    }
    finally {
        cardsCursor.close();
    }
}
```