@marcardar 2016-03-07
You could do, but that's going against the advantages of having the Compat class, which avoids the ugly conditional statements querying the spec version. That can easily get out of control over time, like you hinted at earlier.  Also, asking the question about whether it will be supported by the old provider, is a question future devs will also have, so I think the Compat approach is much cleaner. It's also arguably easier to drop support for older provider versions at a later date.