@timrae 2016-03-07
I think the Compat approach makes sense for the `addNotes()` method, but not really for `addNewNote()` as I feel it's unnecessary obfuscation. I don't really mind if you prefer the Compat approach, but in any case you'll need to get rid of the duplicate code there.