@timrae 2016-03-07
Re: `addNotes()`

I think the DuplicatesStrategy enum idea was probably more functional and general than the code currently is, though I'm still not really sure if that's indeed the best way to handle the issue.

From a high level, I think what we're looking for in this PR is the following:

0. Significant speed improvements from using a single transaction for bulk inserts
0. The ability for an API user to optionally maintain a list of notes (e.g. 'starred' words) and only add the new words without touching the old ones by allowing the skipping of duplicates
0. The ability for an API user to update the fields / tags of existing notes

Anything else?