@timrae 2016-03-07
For point 3 above, I guess the API should probably allow specifying which fields should be updated. AnkiDroid users should be allow to freely modify some fields (to add notes etc) without the host app unnecessarily overwriting them, and they can also append entirely new fields to the model, and that should just work.