@timrae 2016-03-07
I'm thinking maybe the the bulk insert should not handle updates at all, and rather have that handled by an [`applyBatch()`](http://developer.android.com/reference/android/content/ContentResolver.html#applyBatch(java.lang.String, java.util.ArrayList<android.content.ContentProviderOperation>)) implementation, as these are not really strongly coupled. What do you think? I guess if we did this it might be more useful if we could specify which notes were skipped during the insert... That might also be possible with `applyBatch()` but it's hard to say without actually writing the code.