@marcardar 2016-03-07
So you mean bulkInsert() would blindly add, and applyBatch() would do what bulkInsert() does now?