@marcardar 2016-03-07
What I found was that there is never any benefit to adding duplicate notes. It is very easy to do unless the provider (and/or) API protects against it. So the current bulkInsert() implementation does what I think most clients would want without getting into the problems (and potential legacy headaches) that would come with something like DuplicateStrategy. Here we have just one strategy to support in the future.