@marcardar 2016-03-07
Re: your 3 "high level" points:

2. When I add a list of starred words, I want the non-existing ones to be added, and the existing ones to be updated.  This is the current behaviour of addNotes() (and bulkInsert() in spec 2)