@marcardar 2016-03-07
Re: Compat approach. It doesn't make sense to try to share code between different spec versions, because the next spec version may want to use a part of that shared code but not all of it, and so you'll have to replicate anyway. It's really an all or nothing IMHO. It's only about 4 lines anyway or duplicated code anyway, no?