@timrae 2016-03-07
> So you mean bulkInsert() would blindly add, and applyBatch() would do what bulkInsert() does now?

Not quite, I mean that `bulkInsert()` would only support skipping (should probably be default) or not skipping duplicates. Update would then be handled separately by `applyBatch` (since there's no such thing as `bulkUpdate()`).

Then as an after-thought I thought that in the case where we separate these insert and update functionalities, it might be useful for the API users to get told which notes were skipped during the bulk insert. `applyBatch()` can return an array so I was wondering if that could be used *instead* of `bulkInsert()` i.e. to implement both insertion and updating.

Does that make sense? I'm not sure if that would work or if it's a good idea. It's past 3am here in Tokyo so I'd better go to sleep and continue discussing tomorrow :)