@timrae 2016-03-07
> If you want to keep things more simple, we can also drop the addNewNote changes (i.e. get rid of the deckId query parameter support for insert() ). That's a 38% speed improvement which doesn't mean much since those are only called occasionally.

Yeah this is probably for the best, I don't think it's really worth sacrificing readability and/or maintaining support code for such a small improvement, as we can address these performance concerns with the new `addNotes()` method.