@marcardar 2016-03-07
I've never used applyBatch() before, but it looks like a rather hacky, messy solution IMHO.  I just don't think it's worth it for the sake of keeping a bulkInsert() strictly for inserts...