@marcardar 2016-03-08
Here are the timings with the new code:

Old provider (empty deck): export of 2125 words (2111 new) took: 168376ms
New provider (empty deck): export of 2125 words (2111 new) took: 8854ms

Old provider (full deck): export of 2125 words (0 new) took: 747ms
New provider (full deck): export of 2125 words (0 new) took: 678ms